backend/requirements.txt
deepface
flask
numpy
opencv-python-headless
ultralytics
Flask-Cors
pillow
fuzzywuzzy
